rename duckling provider to duckling client
